+++ app/Http/Requests/Device/DevcieTime/UpdateDeviceTimeRequest.php
@@ -44,7 +44,7 @@ class UpdateDeviceTimeRequest extends FormRequest
                         $query->where('device_id', $this->deviceId)
                     ),
             ],
-            'rate'=> ['required','integer','min:1'],
+            'rate'=> ['required','numeric','min:1'],
             'deviceTypeId' => [
                 'nullable',
                 'integer',
